test: organize cressieread python testing, add upper_bound (#3986)
