+++ test/unit_test/distributionally_robust_test.cc
@@ -28,9 +28,7 @@ BOOST_AUTO_TEST_CASE(distributionally_robust_inverse_chisq)
 
 BOOST_AUTO_TEST_CASE(distributionally_robust_recompute_duals)
 {
-  // to generate this data:
-  //
-  // python ./python/tests/DistributionallyRobustUnitTestData.py
+  // To see how this data is generated checkout python/tests/test_distributionall_robust.py
 
   std::pair<double, double> data[] = {{0.4692680899768591, 0.08779271803562538},
       {3.010121430917521, 0.1488852932982503}, {1.3167456935454493, 0.5579699034039329},
@@ -88,9 +86,7 @@ BOOST_AUTO_TEST_CASE(distributionally_robust_recompute_duals)
 
 BOOST_AUTO_TEST_CASE(distributionally_robust_qlb)
 {
-  // to generate this data:
-  //
-  // python ./DistributionallyRobustUnitTestData.py
+  // To see how this data is generated checkout python/tests/test_distributionall_robust.py
 
   std::pair<double, double> data[] = {{0.4692680899768591, 0.08779271803562538},
       {3.010121430917521, 0.1488852932982503}, {1.3167456935454493, 0.5579699034039329},
